# gh-110489: Optimise math.ceil for known exact float