--- a/Misc/NEWS.d/next/Core and Builtins/2023-10-06-22-30-25.gh-issue-110489.rI2n8A.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2023-10-06-22-30-25.gh-issue-110489.rI2n8A.rst
@@ -1,0 +1,1 @@
+Optimise :func:`math.ceil` when the input is exactly a float, resulting in about a 10% improvement.