@hauntsaninja 2023-10-06
Thanks for the review! I'd avoided PEP 7 for consistency with floor, but I just changed both